backend/requirements.txt
flask==2.0.1
google-generativeai==0.3.0
python-dotenv==0.19.0
flask-cors==3.0.10 